Visual Studio 2019 compatibility changes
