+++ imgui.ini
@@ -1,6 +1,6 @@
 [Window][DockspaceViewport_11111111]
 Pos=0,21
-Size=1496964992,452
+Size=32,32
 Collapsed=0
 
 [Window][Debug##Default]
@@ -9,20 +9,20 @@ Size=228,150
 Collapsed=1
 
 [Window][Theme]
-Pos=1194760704,213
-Size=302204288,198
+Pos=26,38
+Size=6,10
 Collapsed=0
 DockId=0x00000007,0
 
 [Window][ImGui Details]
-Pos=1194760704,413
-Size=302204288,60
+Pos=26,50
+Size=6,3
 Collapsed=0
 DockId=0x00000008,0
 
 [Window][Camera Settings]
-Pos=0,293
-Size=283,180
+Pos=0,41
+Size=11,12
 Collapsed=0
 DockId=0x0000000A,0
 
@@ -48,14 +48,14 @@ Size=169,259
 Collapsed=0
 
 [Window][DirectX 12 Settings]
-Pos=1194760704,213
-Size=302204288,198
+Pos=26,38
+Size=6,10
 Collapsed=0
 DockId=0x00000007,1
 
 [Window][Hardware Info]
-Pos=1194760704,21
-Size=302204288,190
+Pos=26,21
+Size=6,15
 Collapsed=0
 DockId=0x00000003,1
 
@@ -78,25 +78,25 @@ Collapsed=0
 DockId=0x00000003,2
 
 [Window][Viewport]
-Pos=285,21
-Size=1194760448,452
+Pos=13,21
+Size=11,32
 Collapsed=0
 DockId=0x00000002,0
 
 [Window][Scene Graph Editor]
 Pos=0,21
-Size=283,270
+Size=11,18
 Collapsed=0
 DockId=0x00000011,0
 
 [Window][Inspector]
-Pos=1194760704,21
-Size=302204288,190
+Pos=26,21
+Size=6,15
 Collapsed=0
 DockId=0x00000003,0
 
 [Docking][Data]
-DockSpace             ID=0x4DDED0D8 Pos=0,21 Size=-8320,452 Split=X
+DockSpace             ID=0x4DDED0D8 Pos=0,21 Size=32,32 Split=X
   DockNode            ID=0x00000004 Parent=0x4DDED0D8 SizeRef=1020,699 Split=X
     DockNode          ID=0x00000001 Parent=0x00000004 SizeRef=283,699 Split=Y SelectedTab=0x1FA76273
       DockNode        ID=0x00000009 Parent=0x00000001 SizeRef=463,418 Split=Y SelectedTab=0x1FA76273
@@ -111,7 +111,7 @@ DockSpace             ID=0x4DDED0D8 Pos=0,21 Size=-8320,452 Split=X
       DockNode        ID=0x0000000A Parent=0x00000001 SizeRef=463,279 SelectedTab=0x8722C6A1
     DockNode          ID=0x00000002 Parent=0x00000004 SizeRef=735,699 CentralNode=1 SelectedTab=0x995B0CF8
   DockNode            ID=0x00000005 Parent=0x4DDED0D8 SizeRef=258,699 Split=Y SelectedTab=0x070A839D
-    DockNode          ID=0x00000003 Parent=0x00000005 SizeRef=236,295 SelectedTab=0x053C54B8
+    DockNode          ID=0x00000003 Parent=0x00000005 SizeRef=236,295 SelectedTab=0xF02CD328
     DockNode          ID=0x00000006 Parent=0x00000005 SizeRef=236,402 Split=Y SelectedTab=0x070A839D
       DockNode        ID=0x00000007 Parent=0x00000006 SizeRef=236,307 SelectedTab=0x070A839D
       DockNode        ID=0x00000008 Parent=0x00000006 SizeRef=236,93 SelectedTab=0x45C34A56
