+++ src/scene_graph/scene_graph.hpp
@@ -4,8 +4,11 @@
 #include <functional>
 #include <memory>
 #include <DirectXMath.h>
+#include <cstdint>
 
-#include "../scene_graph/node.hpp"
+#include "node.hpp"
+#include "light_node.hpp"
+#include "../platform_independend_structs.hpp"
 #include "../util/user_literals.hpp"
 #include "../util/defines.hpp"
 #include "../util/log.hpp"
@@ -24,25 +27,7 @@ namespace wr
 
 	struct CameraNode;
 	struct MeshNode;
-	struct LightNode;
 	struct SkyboxNode;
-
-	enum class LightType : uint32_t
-	{
-		POINT, DIRECTIONAL, SPOT, FREE /* MAX LighType value; but unused */
-	};
-
-	struct Light
-	{
-		DirectX::XMFLOAT3 pos = { 0, 0, 0 };			//Position in world space for spot & point
-		float rad = 5.f;								//Radius for point, height for spot
-
-		DirectX::XMFLOAT3 col = { 1, 1, 1 };			//Color (and strength)
-		uint32_t tid = (uint32_t)LightType::FREE;		//Type id; LightType::x
-
-		DirectX::XMFLOAT3 dir = { 0, 0, 1 };			//Direction for spot & directional
-		float ang = 40._deg;		//Angle for spot; in radians
-	};
 
 	namespace temp {
